apple: Adapter: Apply network settings based on ffi callbacks
The adapter generates the network settings from the tunnel addresseses and the
resources obtained from the ffi callbacks. It saves previously obtained data
for use when it needs to update the settings.

When starting the tunnel, we don't apply the settings until WrappedSession has
started, so that we don't call setTunnelSettings multiple times (once for tunnel
addresses, and once for resources).